sidecar/pyproject.toml
[project]
name = "vanilla-sidecar"
version = "0.1.0"
description = "VanillaDB agent sidecar — FastAPI + CrewAI backend"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "sqlalchemy>=2.0",
    "aiofiles>=23.0",
    "httpx>=0.27",
    "sqlite-vec>=0.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",  # for FastAPI TestClient
]
agents = [
    "crewai>=0.80",
    "litellm>=1.40",
]
ingestion = [
    "marker-pdf>=0.3",
    "pymupdf>=1.24",
    "firecrawl-py>=1.0",
    "crawl4ai>=0.4.0",   # optional: JS-heavy scraping; also run: playwright install chromium
]
mcp = [
    "fastmcp>=0.4",
]
build = [
    "pyinstaller>=6.6",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["../tests/python"]
asyncio_mode = "auto"
